requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.5.2
shellingham==1.5.3
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
